Updated usecase layer to handle product creation

+++ model/product.go
@@ -1,7 +1,7 @@
 package model
 
 type Product struct {
-	ID    int     `json:"id_product"`
+	ID    int64   `json:"id_product"`
 	Name  string  `json:"product_name"`
 	Price float64 `json:"price"`
 }
